scripts/requirements.txt
asgiref==3.6.0
backoff==2.2.1
branca==0.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
folium==0.14.0
idna==3.4
image==1.5.33
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opencage==2.1.1
phonenumbers==8.13.11
Pillow==9.5.0
pycparser==2.21
pyOpenSSL==23.1.1
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
